Cluster Subnetting
This is a rather large commit after a few weeks worth of work trying
to move the cluster to its own subnet that is routable from the host
network into the new cluster subnet.

Currently the subnet lives on 172.17.2.0/24. There are still some hard
coded values that need to be remove from the scripts before pushing
back into main. Additional testing and tear down scripts need to be
added as well. There is a todo list for this that includes:

- Asymetric routing issue
- Hardcoded roles with 172.17.2 in them
- Tear down scripts for the vm host require updating